Window Size Manager + Settings Page Beginnings
Began work on the setting page. Started with creating a global text file that is created if it does not exist when the game is run. Then created a script that can be called in every scene to set the user's default screen preference with two lines! Also created a drop down box to choose what a user wants their default screen setting to be on the settings page.