Systems Node + Default Screen Size + Backgrounds!
Made accessing our systems easier by creating a systems.gd script that can be attatched to a node called 'Systems' in all of our scenes. It will allow us to load all of our systems and scripts there then access them easily in the rest of the scene. Also changed the default height of the project so it was 16:9. Started work on the image system in the display.gd script. As of now we can apply backgrounds to any scene!