Positioning Bug Fixes + Processor Utilization + Face Support + Collision (WIP)
Fixed serious bugs caused by the last commit that added positioning. Removing an image or adding an image could lead to incorrect image placement on the scene. This was fixed. Added a code line to globals.gd that kept the CPU core running the game from jumping to 100% when the game was run. This is apparently something godot does when using their editor by default. Added support for giving faces to bodies so that manipulating the body node always affects the face node as well. Finally I started collision support. Support for meshes was added and body images that use meshes can be added and manipulated like any regular image. There is also now the option to 'slide' a character and avoid any collision when doing so, allowing for moving a character across the screen instead of teleporting them. Actual collision handling has yet to have been implemented.