Basic Collision
Got basic collision working so that when a character collides with another they move until the character they collided with also stops moving. Ideally I would like to improve on it by specifying which nodes should be pushed and not pushed, and a timer for some nodes to continue moving once exited until the timer runs out.