Positioning Bug Fixes + Finished Collision + Fades
The image system is now version 1.0.0. It is possible there is still bugs that will need to be handled once discovered. There was a bug in positioning last commit that broke a node's positioning if the position function was used more than once on it. This caused problems with collision as well. This was fixed. Finished collision by adding a timer and character collide choice to give more control. Finally fades were added for fading in/out from black/alpha. This can be apllied to a single layer or a layer and all its children.