Switch Feature + Index Getter + Additional Face
Realized there were a few things still needed by the image system. This is version 1.1.0. These features were possible before, but not easy to do for the user so I added them to improve the system. The switch feature allows the image texture or video steam of a node to easily be changed. The index getter makes it simple for the user to get the layers[] index of the provided node. Now also allow for one additional face reffered to as 'other'. This other can be a blush, nate's shades, etc.