Volume Feature + Fixed User Settings + Fixed Borderless
Added a volume feature to the game. Includes a master, music, and sfx volume slider. Volume settings are saved in user settings and carry over after the game is closed. Fixed a bug with user settings where if the screen size was not set then the rest of the file was deleted. Fixed a bug with borderless where if the game was started with a default of fullscreen switching to borderless did not make the window the size of the display.