Fade Fixes
Fixed issues with fading where using fade with 9/11 when no body change occured stalled the game. Made it so putting fade in the info[1] spot should never crash the game, that using fade when no pervious body ignores fade altogether, and fixed parsing not being run if body was same and told to fade. I also made it so scene() starts from line 1 instead of 0.