Better Saving + Saving For Fades/Silhouettes
Made saving better by adding 'res://' to character bodies in the character system. This removes the need to check and append 'res://' to bodies that don't have it. Also made saving take fades and silhouettes into account, fixing errors that occured when saving lines that did either. This required a non backwards compatible addition to the save/load system. Thus there is now a check of game version on load. Incompatible saves currently print an error and do nothing.