Below Character AFLs
Support for AFLs below a character. Only works if the character is a systems.display.image() because z_index needed to be used. I attempted to implement a way of adding AFLs below all node types, but when loading from a save it broke the systems.display.nodelayers() function inexplainably. It was like Godot stopped functioning correctly. It was a scarring experience. So that idea was dropped for this.