Fade/Cut To/From Black + Fade Timings + Save Fix
Added script commands to fade to and from black and cut to and from black. Added input for timings on character fades by adding fadeInTime/fadeOutTime after fade in a character info block. Fixed save issue where the modulate of the background was not taken into account when saving as well as crashes when systems.dialogue.lastLayers was [].